requirements.txt
Django>=1.8.0
django-friendship>=1.3.0
